requirements/production.txt
appnope==0.1.0
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
decorator==4.2.1
Django==2.0.3
django-cors-headers==2.2.0
django-extensions==2.0.6
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.6.1
pytz==2018.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
